Remove ominous SourceFileSupport only included in SourceFile
Modules that only define some methods and are only included in
one class are mostly a way to hide that one class is doing too
much. The surface of the resulting class remains the same, just
that it's a bit harder to find out where methods are coming from.

It was/is also unclear when code goes into one file or the other,
at least I don't have a clue how that's working/decided.

Some functions in the main class even went and called functions
in the module direct (process skipped lines for instance).

I think there might be a better separation hiding here, along the
lines of:

* branches vs. lines
* a presenter like object that provides additional convenience for
  formatters (or something like that)
